JVM AOT Cache Demo
Spring Boot 3.3 ships with a new way to extract your application for
production deployment that's also CDS friendly. This commit adds the
dialect our application use to prevent any attempt to open a connection
during the refresh phase.

To use Java AOT Cache (formerly known as CDS) we need to extract our
application to the CDS friendly structure:

java -Djarmode=tools -jar ops-status-0.0.1-SNAPSHOT.jar extract

This creates a directory with the application. To create the CDS cache
we need to run the application with several properties:

* -XX:ArchiveClassesAtExit=application.jsa - to require the JVM to store
the archive on exit in the 'application.jsa' file
* -Dspring.context.exit=onRefresh - to require the Spring Application
context to run in a special mode where it exits once the context has
refreshed. This makes sure the JVM has loaded as many classes as
possible
* -Dspring.flyway.enabled=false - our app uses Flyway on startup and
we don't want it to attempt to upgrade the database as it may not be
available.

Once the file is created, you can use the cache for subsequent run of
the same application:

java -XX:SharedArchiveFile=application.jsa -jar app.jar